Reparaciones Transportista

+++ customers/src/main/java/com/example/dricamp/controllers/ParadaController.java
@@ -1,6 +1,11 @@
 package com.example.dricamp.controllers;
 
+import com.example.dricamp.models.UsuarioRol;
+import com.example.dricamp.models.Vehiculo;
+import io.swagger.v3.oas.annotations.Operation;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,31 +16,58 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/paradas")
+@RequestMapping("/api/paradas")
 public class ParadaController {
 
     @Autowired
     private ParadaService paradaService;
 
+    @Operation(summary = "Obtiene todos las paradas")
     @GetMapping
     public List<Parada> getAllParadas() {
         return paradaService.getAllParadas();
     }
 
+    @Operation(summary = "Obtiene una parada por su ID")
     @GetMapping("/{id}")
-    public ResponseEntity<Parada> getParadaById(@PathVariable Long id) {
-        Optional<Parada> parada = paradaService.getParadaById(id);
-        return parada.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+    public Parada getParadaById(@PathVariable Long id) {
+        return paradaService.getParadaById(id);
     }
 
+    @Operation(summary = "Realiza la creación de una parada")
     @PostMapping
-    public Parada saveParada(@RequestBody Parada parada) {
-        return paradaService.saveParada(parada);
+    public ResponseEntity<Parada> saveParada(@RequestBody Parada parada) {
+        Parada obj = paradaService.saveParada(parada);
+        return new ResponseEntity<>(obj, HttpStatus.OK);
     }
 
+    @Operation(summary = "Elimina un parada")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteParada(@PathVariable Long id) {
-        paradaService.deleteParada(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Parada> deleteParada(@PathVariable Long id) {
+        Parada obj = paradaService.getParadaById(id);
+        if(obj != null){
+            paradaService.deleteParada(id);
+        }else{
+            return new ResponseEntity<>(obj, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>(obj, HttpStatus.OK);
     }
+
+    @Operation(summary = "Modifica una parada")
+    @PutMapping
+    public ResponseEntity<Parada> update(@Valid @RequestBody Parada parada){
+        Parada obj = paradaService.getParadaById(parada.getIdParada());
+        if (obj != null){
+            obj.setIdParada(parada.getIdParada());
+            obj.setDepartamento(parada.getDepartamento());
+            obj.setCiudad(parada.getCiudad());
+            obj.setEstacion(parada.getEstacion());
+            paradaService.saveParada(obj);
+
+        }else{
+            return new ResponseEntity<>(obj, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>(obj, HttpStatus.OK);
+    }
+
 }
